.claude/requirements.txt
# AI layer dependencies — advisory system (external LLM providers)
# Installed by init.sh into the project venv
# App dependencies go in pyproject.toml, not here
openai
google-genai
